Space/requirements.txt
# SolarVision AI - Hugging Face Spaces Requirements
# Install with: pip install -r requirements.txt

# Core ML
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Web Framework
gradio>=4.0.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data & Visualization
pandas>=2.0.0
matplotlib>=3.7.0

# PDF Generation
reportlab>=3.6.0

# Utilities
tqdm>=4.66.0
